refactor: WinnerDecision 폴더구조 변경
