+++ src/main/java/racingcar/util/WinnerDecision.java
@@ -1,7 +1,8 @@
-package racingcar.model;
+package racingcar.util;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.model.Car;
 
 public class WinnerDecision {
